Allows the execution of only one job at once.
Set
spec:
  concurrencyPolicy: Forbid

in values.yaml in order to run only one job at once (the successive job will be
queued and executed at the end of the running one).

Set
spec:
  concurrencyPolicy: Forbid
  startingDeadlineSeconds: 10

in values.yaml in order to run only one job at once and skip concurrent ones
(successive jobs will not be queued).

The label cron-job-semaphore is set in order to be used as semaphore by the
running application.